endpoint that can be called to repair dataset names
